# Improve the performance of removing variables from expressions