--- a/dimod/include/dimod/expression.h
+++ b/dimod/include/dimod/expression.h
@@ -419,7 +419,20 @@ void Expression<bias_type, index_type>::clear() {
 template <class bias_type, class index_type>
 template <class T>
 void Expression<bias_type, index_type>::fix_variable(index_type v, T assignment) {
-    throw std::logic_error("not implemented - fix_variable");
+    assert(v >= 0 && static_cast<size_type>(v) < parent_->num_variables());
+
+    auto vit = indices_.find(v);
+    if (vit == indices_.end()) return;  // nothing to remove
+
+    // remove the biases
+    base_type::fix_variable(vit->second, assignment);
+
+    // update the indices
+    auto it = variables_.erase(variables_.begin() + vit->second);
+    indices_.erase(vit);
+    for (; it != variables_.end(); ++it) {
+        indices_[*it] -= 1;
+    }
 }
 
 template <class bias_type, class index_type>
@@ -592,6 +605,8 @@ bool Expression<bias_type, index_type>::remove_interaction(index_type u, index_t
 
 template <class bias_type, class index_type>
 void Expression<bias_type, index_type>::remove_variable(index_type v) {
+    assert(v >= 0 && static_cast<size_type>(v) < parent_->num_variables());
+
     auto vit = indices_.find(v);
     if (vit == indices_.end()) return;  // nothing to remove
 
@@ -599,12 +614,10 @@ void Expression<bias_type, index_type>::remove_variable(index_type v) {
     base_type::remove_variable(vit->second);
 
     // update the indices
-    variables_.erase(variables_.begin() + vit->second);
-
-    // indices is no longer valid, so remake
-    indices_.clear();
-    for (size_type ui = 0; ui < variables_.size(); ++ui) {
-        indices_[variables_[ui]] = ui;
+    auto it = variables_.erase(variables_.begin() + vit->second);
+    indices_.erase(vit);
+    for (; it != variables_.end(); ++it) {
+        indices_[*it] -= 1;
     }
 }