--- a/dimod/include/dimod/quadratic_model.h
+++ b/dimod/include/dimod/quadratic_model.h
@@ -64,6 +64,15 @@ class QuadraticModel : public abc::QuadraticModelBase<Bias, Index> {
     /// Change the vartype of `v`, updating the biases appropriately.
     void change_vartype(Vartype vartype, index_type v);
 
+    /**
+     * Remove variable `v` from the model by fixing its value.
+     *
+     * Note that this causes a reindexing, where all variables above `v` have
+     * their index reduced by one.
+     */
+    template <class T>
+    void fix_variable(index_type v, T assignment);
+
     /// Return the lower bound on variable ``v``.
     bias_type lower_bound(index_type v) const;
 
@@ -227,6 +236,13 @@ void QuadraticModel<bias_type, index_type>::change_vartype(Vartype vartype, inde
         // todo: there are more we could support
         throw std::logic_error("unsupported vartype change");
     }
+}
+
+template <class bias_type, class index_type>
+template <class T>
+void QuadraticModel<bias_type, index_type>::fix_variable(index_type v, T assignment) {
+    base_type::fix_variable(v, assignment);
+    varinfo_.erase(varinfo_.begin() + v);
 }
 
 template <class bias_type, class index_type>