This provides a pretty significant performance improvement when fixing variables in individual constraints.
```python
import time

import dimod

num_variables = 10_000

cqm = dimod.ConstrainedQuadraticModel()
cqm.add_variables("BINARY", num_variables)
c0 = cqm.add_constraint(((v, 1) for v in range(num_variables)), '==', 1)
expr = cqm.constraints[c0].lhs

t = time.perf_counter()
for v in range(num_variables):
    expr.fix_variable(v, 0)
print(time.perf_counter() - t)
```
Gives a time of ~0.15s vs 1.37s on 0.12.4 on my system.

However, this only makes a very minor improvement when fixing variables in the overall model (see https://github.com/dwavesystems/dimod/issues/1317)
```python
import time

import dimod

num_variables = 10_000

cqm = dimod.ConstrainedQuadraticModel()
cqm.add_variables("BINARY", num_variables)
c0 = cqm.add_constraint(((v, 1) for v in range(num_variables)), '==', 1)

t = time.perf_counter()
for v in range(num_variables):
    cqm.fix_variable(v, 0)
print(time.perf_counter() - t)
```
Takes ~25s vs ~28s on 0.12.4 on my system.